requirements.txt
aiomysql==0.1.1 ; python_version >= "3.9" and python_version < "3.12"
argcomplete==2.0.0 ; python_version >= "3.9" and python_version < "3.12"
asn1crypto==1.5.1 ; python_version >= "3.9" and python_version < "3.12"
attrs==22.2.0 ; python_version >= "3.9" and python_version < "3.12"
boto3==1.26.74 ; python_version >= "3.9" and python_version < "3.12"
botocore==1.29.74 ; python_version >= "3.9" and python_version < "3.12"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "3.12"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "3.12"
cfgv==3.3.1 ; python_version >= "3.9" and python_version < "3.12"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "3.12"
click==8.1.3 ; python_version >= "3.9" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "3.12"
colorlog==6.7.0 ; python_version >= "3.9" and python_version < "3.12"
commonmark==0.9.1 ; python_version >= "3.9" and python_version < "3.12"
coverage[toml]==7.1.0 ; python_version >= "3.9" and python_version < "3.12"
cryptography==39.0.1 ; python_version >= "3.9" and python_version < "3.12"
databases[mysql]==0.7.0 ; python_version >= "3.9" and python_version < "3.12"
distlib==0.3.6 ; python_version >= "3.9" and python_version < "3.12"
exceptiongroup==1.1.0 ; python_version >= "3.9" and python_version < "3.11"
filelock==3.9.0 ; python_version >= "3.9" and python_version < "3.12"
greenlet==2.0.2 ; python_version >= "3.9" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "3.12"
hypothesis==6.68.2 ; python_version >= "3.9" and python_version < "3.12"
identify==2.5.18 ; python_version >= "3.9" and python_version < "3.12"
idna==3.4 ; python_version >= "3.9" and python_version < "3.12"
iniconfig==2.0.0 ; python_version >= "3.9" and python_version < "3.12"
jellyfish==0.9.0 ; python_version >= "3.9" and python_version < "3.12"
jmespath==1.0.1 ; python_version >= "3.9" and python_version < "3.12"
joblib==1.2.0 ; python_version >= "3.9" and python_version < "3.12"
mypy-extensions==1.0.0 ; python_version >= "3.9" and python_version < "3.12"
mypy==1.0.1 ; python_version >= "3.9" and python_version < "3.12"
nodeenv==1.7.0 ; python_version >= "3.9" and python_version < "3.12"
nox-poetry==1.0.2 ; python_version >= "3.9" and python_version < "3.12"
nox==2022.11.21 ; python_version >= "3.9" and python_version < "3.12"
numpy==1.24.2 ; python_version < "3.12" and python_version >= "3.9"
oscrypto==1.3.0 ; python_version >= "3.9" and python_version < "3.12"
packaging==23.0 ; python_version >= "3.9" and python_version < "3.12"
pandas==1.5.3 ; python_version >= "3.9" and python_version < "3.12"
platformdirs==3.0.0 ; python_version >= "3.9" and python_version < "3.12"
pluggy==1.0.0 ; python_version >= "3.9" and python_version < "3.12"
pre-commit==3.0.4 ; python_version >= "3.9" and python_version < "3.12"
pyarrow==10.0.1 ; python_version >= "3.9" and python_version < "3.12"
pycparser==2.21 ; python_version >= "3.9" and python_version < "3.12"
pycryptodomex==3.17 ; python_version >= "3.9" and python_version < "3.12"
pydantic[dotenv]==1.10.5 ; python_version >= "3.9" and python_version < "3.12"
pygments==2.14.0 ; python_version >= "3.9" and python_version < "3.12"
pyjwt==2.6.0 ; python_version >= "3.9" and python_version < "3.12"
pymysql==1.0.2 ; python_version >= "3.9" and python_version < "3.12"
pyopenssl==23.0.0 ; python_version >= "3.9" and python_version < "3.12"
pytest-cov==4.0.0 ; python_version >= "3.9" and python_version < "3.12"
pytest-mock==3.10.0 ; python_version >= "3.9" and python_version < "3.12"
pytest==7.2.1 ; python_version >= "3.9" and python_version < "3.12"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.12"
python-dotenv==0.21.1 ; python_version >= "3.9" and python_version < "3.12"
pytz==2022.7.1 ; python_version >= "3.9" and python_version < "3.12"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "3.12"
requests==2.28.2 ; python_version >= "3.9" and python_version < "3.12"
rich==12.6.0 ; python_version >= "3.9" and python_version < "3.12"
s3transfer==0.6.0 ; python_version >= "3.9" and python_version < "3.12"
scikit-learn==1.2.1 ; python_version >= "3.9" and python_version < "3.12"
scipy==1.10.0 ; python_version >= "3.9" and python_version < "3.12"
setuptools==67.3.2 ; python_version >= "3.9" and python_version < "3.12"
shellingham==1.5.0.post1 ; python_version >= "3.9" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.12"
snowflake-connector-python[pandas]==3.0.0 ; python_version >= "3.9" and python_version < "3.12"
sortedcontainers==2.4.0 ; python_version >= "3.9" and python_version < "3.12"
sqlalchemy==1.4.46 ; python_version >= "3.9" and python_version < "3.12"
structlog==22.3.0 ; python_version >= "3.9" and python_version < "3.12"
threadpoolctl==3.1.0 ; python_version >= "3.9" and python_version < "3.12"
tomli==2.0.1 ; python_version >= "3.9" and python_full_version <= "3.11.0a6"
tomlkit==0.11.6 ; python_version >= "3.9" and python_version < "3.12"
typer[all]==0.7.0 ; python_version >= "3.9" and python_version < "3.12"
types-python-dateutil==2.8.19.7 ; python_version >= "3.9" and python_version < "3.12"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "3.12"
urllib3==1.26.14 ; python_version >= "3.9" and python_version < "3.12"
virtualenv==20.19.0 ; python_version >= "3.9" and python_version < "3.12"
